Move the abstract MentionType classes up one level for consistency sake
