+++ sources/ElkArte/Mentions/MentionType/Event/Mailfail.php
@@ -13,12 +13,14 @@
 
 namespace ElkArte\Mentions\MentionType\Event;
 
+use ElkArte\Mentions\MentionType\AbstractEventBoardAccess;
+
 /**
  * Class MailfailMention
  *
  * Handles notifying users who have had email notifications disabled for failure to deliver
  */
-class Mailfail extends AbstractMentionBoardAccess
+class Mailfail extends AbstractEventBoardAccess
 {
 	/**
 	 * {@inheritdoc }
