+++ sources/ElkArte/Mentions/MentionType/Event/Mentionmem.php
@@ -13,6 +13,7 @@
 
 namespace ElkArte\Mentions\MentionType\Event;
 
+use ElkArte\Mentions\MentionType\AbstractEventBoardAccess;
 use ElkArte\DataValidator;
 use ElkArte\Mentions\Mentioning;
 use ElkArte\Notifications;
@@ -23,7 +24,7 @@ use ElkArte\NotificationsTask;
  *
  * Handles the mentioning of members (@ member actions)
  */
-class Mentionmem extends AbstractMentionBoardAccess
+class Mentionmem extends AbstractEventBoardAccess
 {
 	/**
 	 * {@inheritdoc }
