+++ sources/ElkArte/Mentions/MentionType/Event/Quotedmem.php
@@ -13,6 +13,7 @@
 
 namespace ElkArte\Mentions\MentionType\Event;
 
+use ElkArte\Mentions\MentionType\AbstractEventBoardAccess;
 use ElkArte\DataValidator;
 use ElkArte\Mentions\Mentioning;
 use ElkArte\Notifications;
@@ -23,7 +24,7 @@ use ElkArte\NotificationsTask;
  *
  * Handles mentioning of members whose messages has been quoted
  */
-class Quotedmem extends AbstractMentionBoardAccess
+class Quotedmem extends AbstractEventBoardAccess
 {
 	/**
 	 * {@inheritdoc }
