+++ sources/ElkArte/Mentions/MentionType/Notification/Mailfail.php
@@ -13,12 +13,14 @@
 
 namespace ElkArte\Mentions\MentionType\Notification;
 
+use ElkArte\Mentions\MentionType\AbstractNotificationBoardAccess;
+
 /**
  * Class MailfailMention
  *
  * Handles notifying users who have had email notifications disabled for failure to deliver
  */
-class Mailfail extends AbstractMentionBoardAccess
+class Mailfail extends AbstractNotificationBoardAccess
 {
 	/**
 	 * {@inheritdoc }
