+++ sources/ElkArte/Mentions/MentionType/Notification/Mentionmem.php
@@ -13,12 +13,14 @@
 
 namespace ElkArte\Mentions\MentionType\Notification;
 
+use ElkArte\Mentions\MentionType\AbstractNotificationBoardAccess;
+
 /**
  * Class MentionmemMention
  *
  * Handles the mentioning of members (@ member actions)
  */
-class Mentionmem extends AbstractMentionBoardAccess
+class Mentionmem extends AbstractNotificationBoardAccess
 {
 	/**
 	 * {@inheritdoc }
